+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Under development
 -----------------
+- Feature: Add `summary` and `defaultValue` for InputTag (alexeevdv)
 
 1.1.0 2019-04-12
 -----------------